+++ test/unit/property.js
@@ -34,23 +34,18 @@ describe('swiftserver:property', function() {
 
     var runContext;
     before(function() {
-      // Mock the options, set up an output folder and run the generator
       runContext = helpers.run(path.join( __dirname, '../../property'))
         .withGenerators(dependentGenerators) // Stub subgenerators
-          //Get the temporary directory so that we can delete later on.
         .inTmpDir(function(tmpDir){
-
           var tmpFile = path.join(tmpDir, ".swiftservergenerator-project");    //Created to make the dir a kitura project
           fs.writeFileSync(tmpFile, "");
-
-          var config = {
+          var spec = {
             appName: "modelTestDir",
-            store: "cloudant",
-            logger: "helium",
-            port: 8090
+            appType: "crud",
+            config: {}
           }
-          var tmpConfig = path.join(tmpDir, "config.json");           //Created so there is an appName for the models
-          fs.writeFileSync(tmpConfig, JSON.stringify(config));
+          var tmpSpec = path.join(tmpDir, "spec.json");           //Created so there is an appName for the models
+          fs.writeFileSync(tmpSpec, JSON.stringify(spec));
 
           var tmp_yorc = path.join(tmpDir, ".yo-rc.json");             //Created so we can test in a different directory
           fs.writeFileSync(tmp_yorc, "{}");
@@ -70,7 +65,7 @@ describe('swiftserver:property', function() {
         })
         .withPrompts({                       // Mock the prompt answers
           model: 'MyModel',
-          name: 'foo',
+          propertyName: 'foo',
           type: 'boolean'
         })
         return runContext.toPromise();                        // Get a Promise back when the generator finishes